Packaged class utilities and contexts
In order to get these out of the init files as much as possible, started down a
path of reorganization in which there are more smaller and single-purpose
modules.
