+++ salve/__init__.py
@@ -1,9 +1,10 @@
+from salve.class_resources import Enum, Singleton, with_metaclass
+
 import salve.log
 import salve.context
 
 __version__ = '2.3.1'
 
-exec_context = salve.context.ExecutionContext(
-    startphase=salve.context.ExecutionContext.phases.STARTUP)
+exec_context = salve.context.ExecutionContext()
 
 logger = salve.log.Logger(exec_context)
