+++ salve/action/create.py
@@ -4,10 +4,9 @@ import abc
 
 import salve
 
-from salve import action, paths
+from salve import action, paths, with_metaclass
 from salve.filesys import access_codes
 from salve.context import ExecutionContext
-from salve.util import with_metaclass
 
 
 class CreateAction(with_metaclass(abc.ABCMeta, action.Action)):
