+++ salve/class_resources/__init__.py
@@ -1,0 +1,7 @@
+# This is a package providing shared resources for the python code itself
+# pertaining to class structure, heirarchy, metaclassing, &c
+
+from salve.class_resources.enum import Enum
+from salve.class_resources.singleton import Singleton
+
+from salve.class_resources.metaclass import with_metaclass
