+++ salve/filesys/__init__.py
@@ -2,7 +2,7 @@
 
 import os
 
-from salve.util import Enum
+from salve import Enum
 from salve.filesys.abstract import Filesys
 from salve.filesys.concrete import Filesys as ConcreteFilesys
